requirements.txt
streamlit==1.32.0
plotly==5.18.0
pandas==2.2.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
nltk>=3.8.0
tqdm>=4.65.0
google-generativeai==0.3.2
python-dotenv==1.0.1
flask==3.0.2
gunicorn==21.2.0
prometheus-client==0.20.0
werkzeug==3.0.1 